refactor: implementing new error parser
